Add support for large messages
Rather than having to verify a whole memory buffer in one go, allow chunks of
memory to be processed using begin, update and end verify functions. The begin
function starts a new verify process, returning an opaque context handle. The
update function can then be called multiple times with this context and blocks
of memory to process memory in chuncks. Finally the end function is called with
the context to end the verify process and get the result.